November 10, 2025: hello CSS battle
My first battle! I was feeling so very overwhelmed by
this and I probably could have used a simpler dom for
this, but I was pleasantly surprised at how satisfying
it was to build simple (lol) CSS layouts from nothing
and just have a very basic goal in mind. I wonder if
I'll stick with it?

Accuracy was horrendously low because I changed the parent
div's margins/width/height at the last minute and didn't
think about the implications

Accuracy: 88.5%